Completely revised PerfectGases package.
This is for #19.
Functions that are not declared in the base class are now made
protected. This is to avoid that models use them, and then become
incompatible with other media models that do not implement these
functions.
The order of functions has been changed to be alphabetically.
MoistAirUnsaturated now contains all implementations, as this is the
medium that is typically used.
MoistAir extends from MoistAirUnsaturated, and only implements what is
different.
Comments have been updated throughout the packages.