Moved IdealGases.SimpleAir to PerfectGases.DryAir
This is for #19.
The reason for moving this medium is to have a consistent naming. Now,
we have MoistAir(Unsaturated) and DryAir, as opposed to
MoistAir(Unsaturated) and SimpleAir.
The medium IdealGases.SimpleAir is also a perfect gas, and all
Media.PerfectGases are ideal gases. Hence, it makes sense to put them
in the same package.