Add 2 examples to functions #2
This commit adds examples for isMonotonic and trapezoidalIntegration
functions. The example model for trapezoidalIntegration works with a
integrands consisting only of integer values. For another set of
integrands of format Real the numerical error seems to be quite large.
This second example is included in the model but commented out.