Exposed parameter initialize_p.
This is required to be able to overwrite the default value of this
parameter.
The default value needs to be overwritten for models that stagger
multiple volumes, such as a finite volume model of a heat exchanger. In
this situation, all except the first volume need to set initialize_p =
false.