Corrected inflection point in dp/dr for #458
This refactores the pump curves to avoid a numerical problem reported by IDEAS. See the top of Movers.UserGuide for an explanation of the root cause. This change is not backwards compatible as some functions in Characteristics have been deleted and for some the argument list has changed. However, users probably never call these functions directly. The reference results have been updated and generated with Dymola 2017